requirements.txt
kafka-python
faker
psycopg2-binary
streamlit-folium
pandas
plotly
folium 